refactor: 관심사에 맞게 유효성 검사 분리
